添加 MapStruct mapper和 底层Detail接口参数为interger数据库是Long报错

+++ src/main/java/com/letgym/gymcourse/core/AbstractService.java
@@ -44,7 +44,7 @@ public abstract class AbstractService<T> implements Service<T> {
         mapper.updateByPrimaryKeySelective(model);
     }
 
-    public T findById(Integer id) {
+    public T findById(Long id) {
         return mapper.selectByPrimaryKey(id);
     }
